requirements.txt
django
djangorestframework
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
langchain-groq
resend
